refactor(ui): extract finding group resource state into shared hooks
Replace the component-local state and drawer wiring duplicated by findings-group-drill-down and inline-resource-container with two shared hooks: useFindingGroupResources (renamed from useInfiniteResources) handles infinite scroll fetching, and useFindingGroupResourceState composes it with row selection and the resource detail drawer.
