+++ ui/hooks/use-finding-group-resources.ts
@@ -12,7 +12,7 @@ import { FindingResourceRow } from "@/types";
 
 const RESOURCES_PAGE_SIZE = 10;
 
-interface UseInfiniteResourcesOptions {
+interface UseFindingGroupResourcesOptions {
   checkId: string;
   hasDateOrScanFilter: boolean;
   filters: Record<string, string | string[] | undefined>;
@@ -22,28 +22,17 @@ interface UseInfiniteResourcesOptions {
     hasMore: boolean,
   ) => void;
   onSetLoading: (loading: boolean) => void;
-  /** Scroll container element for IntersectionObserver root. Defaults to viewport. */
   scrollContainerRef?: React.RefObject<HTMLElement | null>;
 }
 
-interface UseInfiniteResourcesReturn {
+interface UseFindingGroupResourcesReturn {
   sentinelRef: (node: HTMLDivElement | null) => void;
-  /** Reset pagination and re-fetch page 1 (e.g. after muting). */
   refresh: () => void;
-  /** Imperatively load the next page (e.g. from drawer navigation). */
   loadMore: () => void;
-  /** Total number of resources matching current filters (from API pagination). */
   totalCount: number | null;
 }
 
-/**
- * Hook for paginated infinite-scroll loading of finding group resources.
- *
- * Uses refs for all mutable state to avoid dependency chains that
- * cause infinite re-render loops. The parent component remounts this
- * hook via key-prop when checkId or filters change.
- */
-export function useInfiniteResources({
+export function useFindingGroupResources({
   checkId,
   hasDateOrScanFilter,
   filters,
@@ -51,28 +40,21 @@ export function useInfiniteResources({
   onAppendResources,
   onSetLoading,
   scrollContainerRef,
-}: UseInfiniteResourcesOptions): UseInfiniteResourcesReturn {
-  // All mutable state in refs to break dependency chains
+}: UseFindingGroupResourcesOptions): UseFindingGroupResourcesReturn {
   const pageRef = useRef(1);
   const hasMoreRef = useRef(true);
-  // Start as `true` to block the IntersectionObserver from calling loadNextPage
-  // before the initial fetch runs. Ref callbacks fire during commit (sync),
-  // but useMountEffect fires after paint — the observer can sneak in between.
   const isLoadingRef = useRef(true);
   const currentCheckIdRef = useRef(checkId);
   const controllerRef = useRef<AbortController | null>(null);
   const observerRef = useRef<IntersectionObserver | null>(null);
   const totalCountRef = useRef<number | null>(null);
 
-  // Store latest values in refs so the fetch function always reads current values
-  // without being recreated on every render
   const hasDateOrScanRef = useRef(hasDateOrScanFilter);
   const filtersRef = useRef(filters);
   const onSetResourcesRef = useRef(onSetResources);
   const onAppendResourcesRef = useRef(onAppendResources);
   const onSetLoadingRef = useRef(onSetLoading);
 
-  // Keep refs in sync with latest props
   currentCheckIdRef.current = checkId;
   hasDateOrScanRef.current = hasDateOrScanFilter;
   filtersRef.current = filters;
@@ -103,7 +85,6 @@ export function useInfiniteResources({
         filters: filtersRef.current,
       });
 
-      // Discard stale response if aborted (e.g. Strict Mode remount)
       if (signal.aborted) {
         return;
       }
@@ -116,10 +97,6 @@ export function useInfiniteResources({
       const hasMore = page < totalPages;
       totalCountRef.current = response?.meta?.pagination?.count ?? null;
 
-      // Commit the page number only after a successful (non-aborted) fetch.
-      // This prevents a premature pageRef increment from loadNextPage being
-      // permanently committed if a concurrent abort fires before fetchPage
-      // starts executing.
       pageRef.current = page;
       hasMoreRef.current = hasMore;
 
@@ -130,27 +107,20 @@ export function useInfiniteResources({
       }
     } catch (error) {
       if (!signal.aborted) {
-        console.error("Error fetching resources:", error);
+        console.error("Error fetching finding group resources:", error);
         onSetLoadingRef.current(false);
       }
     } finally {
-      // Only release the loading guard if this fetch wasn't aborted.
-      // An aborted fetch (e.g. Strict Mode cleanup) must NOT reset the flag
-      // while a subsequent fetch from the remount is still in flight.
       if (!signal.aborted) {
         isLoadingRef.current = false;
       }
     }
   }
 
-  // Fetch first page on mount — parent remounts via key-prop on checkId/filter changes
   useMountEffect(() => {
     const controller = new AbortController();
     controllerRef.current = controller;
 
-    // Release the loading guard so fetchPage can proceed.
-    // This is synchronous with the fetchPage call below, so the observer
-    // cannot sneak in between these two lines.
     isLoadingRef.current = false;
     fetchPage(1, false, checkId, controller.signal);
 
@@ -167,17 +137,13 @@ export function useInfiniteResources({
       isLoadingRef.current ||
       !signal ||
       signal.aborted
-    )
+    ) {
       return;
+    }
 
-    // Pass the next page number as an argument without pre-committing
-    // pageRef.current. The fetchPage function commits pageRef.current = page
-    // only after a successful (non-aborted) response, eliminating the race
-    // where a concurrent abort would leave pageRef permanently incremented.
     fetchPage(pageRef.current + 1, true, currentCheckIdRef.current, signal);
   }
 
-  // IntersectionObserver callback
   function handleIntersection(entries: IntersectionObserverEntry[]) {
     const [entry] = entries;
     if (entry.isIntersecting) {
@@ -185,7 +151,6 @@ export function useInfiniteResources({
     }
   }
 
-  // Set up observer when sentinel node changes
   function sentinelRef(node: HTMLDivElement | null) {
     if (observerRef.current) {
       observerRef.current.disconnect();
@@ -201,7 +166,6 @@ export function useInfiniteResources({
     }
   }
 
-  /** Imperatively reset and re-fetch page 1 without changing deps. */
   function refresh() {
     controllerRef.current?.abort();
     const controller = new AbortController();
